// src/generated/models/Cr9b5_forecastpropertiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr9b5_forecastpropertiesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr9b5_forecastpropertiesstatecode = keyof typeof Cr9b5_forecastpropertiesstatecode;
export const Cr9b5_forecastpropertiesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr9b5_forecastpropertiesstatuscode = keyof typeof Cr9b5_forecastpropertiesstatuscode;

export interface Cr9b5_forecastpropertiesBase {
  "cr9b5_forecastflowid@odata.bind"?: string;
  cr9b5_forecastpropertyid: string;
  cr9b5_name?: string;
  "cr9b5_propertyid@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr9b5_forecastpropertiesstatecode;
  statuscode?: Cr9b5_forecastpropertiesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr9b5_forecastproperties extends Cr9b5_forecastpropertiesBase {
  cr9b5_forecastflowidname?: string;
  cr9b5_propertyidname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr9b5_forecastflowid?: object;
  _cr9b5_forecastflowid_value?: string;
  cr9b5_propertyid?: object;
  _cr9b5_propertyid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}